+++ src/lib/pages/add-network-manual/index.tsx
@@ -2,10 +2,12 @@ import { Flex } from "@chakra-ui/react";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 
+import { INITIA_DECODER } from "config/chain/initia";
 import ActionPageContainer from "lib/components/ActionPageContainer";
 import { CustomIcon } from "lib/components/icon";
 import { FooterCta } from "lib/components/layouts";
 import { CelatoneSeo } from "lib/components/Seo";
+import { useChainConfigStore } from "lib/providers/store";
 
 import {
   AddNetworkStepper,
@@ -24,6 +26,8 @@ import {
 import type { AddNetworkManualForm } from "./types";
 
 export const AddNetworkManual = () => {
+  const { addChainConfig } = useChainConfigStore();
+
   const {
     control,
     handleSubmit,
@@ -88,9 +92,93 @@ export const AddNetworkManual = () => {
     assets,
   } = watch();
 
-  const handleSubmitForm = () => {
-    // alert("Please view the console to see the form data");
-    // console.log(JSON.stringify(data, null, 2));
+  const handleSubmitForm = (data: AddNetworkManualForm) => {
+    addChainConfig(data.chainId, {
+      tier: "sequencer",
+      chainId: data.chainId,
+      chain: "initia",
+      registryChainName: data.registryChainName,
+      prettyName: data.networkName,
+      logo_URIs: {
+        png: data.logoUri,
+      },
+      lcd: data.lcdUrl,
+      rpc: data.rpcUrl,
+      graphql: "",
+      wallets: ["initia", "keplr"],
+      features: {
+        faucet: {
+          enabled: false,
+        },
+        wasm: data.isWasm
+          ? {
+              enabled: true,
+              storeCodeMaxFileSize: 1024 * 1024 * 2,
+              clearAdminGas: 1000000,
+            }
+          : { enabled: false },
+        move: data.isMove
+          ? {
+              enabled: true,
+              moduleMaxFileSize: 1_048_576,
+              decodeApi: INITIA_DECODER,
+              verify: "",
+            }
+          : { enabled: false },
+        pool: {
+          enabled: false,
+        },
+        publicProject: {
+          enabled: true,
+        },
+        gov: {
+          enabled: false,
+        },
+        nft: {
+          enabled: data.isNfts,
+        },
+      },
+      gas: {
+        gasAdjustment: Number(data.gasAdjustment),
+        maxGasLimit: Number(data.maxGasLimit),
+      },
+      extra: {
+        isValidatorExternalLink: null,
+        layer: "2",
+      },
+      network_type: "testnet",
+      fees: {
+        fee_tokens: [
+          {
+            denom: data.feeTokenDenom,
+            fixed_min_gas_price: Number(data.fixedMinimumGasPrice),
+            low_gas_price: Number(data.lowGasPrice),
+            average_gas_price: Number(data.averageGasPrice),
+            gas_costs: {
+              cosmos_send: Number(data.gasForCosmosSend),
+              ibc_transfer: Number(data.gasForIbc),
+            },
+          },
+        ],
+      },
+      registry: {
+        bech32_prefix: data.bech32Prefix,
+        slip44: data.slip44,
+        staking: {
+          staking_tokens: [],
+        },
+        assets: data.assets.map((asset) => ({
+          name: asset.name,
+          base: asset.base,
+          symbol: asset.symbol,
+          denom_units: asset.denoms.map((denom) => ({
+            denom: denom.denom,
+            exponent: Number(denom.exponent),
+          })),
+          display: asset.symbol,
+        })),
+      },
+    });
   };
 
   const { currentStep, handleNext, handlePrevious, hasNext, hasPrevious } =