+++ src/pages/_app.tsx
@@ -27,6 +27,7 @@ const isBrowser = typeof window !== "undefined";
 configurePersistable({
   storage: isBrowser ? localforage : undefined,
   stringify: false,
+  debugMode: true,
 });
 
 const MyApp = ({ Component, pageProps }: AppProps) => (